moving search index schema to spyglass-searcher crate
