+++ crates/entities/src/lib.rs
@@ -1,6 +1,5 @@
 use std::collections::HashMap;
 pub mod models;
-pub mod schema;
 pub mod test;
 
 pub use sea_orm;
